Time: 3 ms (87.16%) | Memory: 72 MB (24.67%) - LeetSync

+++ 1445-number-of-sub-arrays-of-size-k-and-average-greater-than-or-equal-to-threshold/number-of-sub-arrays-of-size-k-and-average-greater-than-or-equal-to-threshold.java
@@ -6,11 +6,11 @@ class Solution {
         for(int i=0;i<k;i++){
             sum+=arr[i];
         }
-            avg=(sum)/k;
+             avg=(double)sum/k;
             if(avg>=threshold) count++;
         for(int i=k;i<arr.length;i++){
             sum+=arr[i]-arr[i-k];
-            avg=(sum)/k;
+             avg=(double)sum/k;
             if(avg>=threshold) count++;
         }
         return count;
